requirements.txt
numpy
scipy
qiskit
perceval-quandela
qiskit_aer
qiskit_ibm_runtime